requirements.txt
speedtest-cli
ping3
customtkinter